# swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "title": "MERN petslove",
    "description": "Description",
    "version": "1.0.0"
  },
  "host": "localhost:8080",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "paths": {
    "/api/users/signup": {
      "post": {
        "tags": [
          "User-SignUp & Login/out"
        ],
        "description": "validate",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "username": {
                  "example": "any"
                },
                "password": {
                  "example": "any"
                },
                "confirmPassword": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/users/login": {
      "post": {
        "tags": [
          "User-SignUp & Login/out"
        ],
        "description": "validate",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "username": {
                  "example": "any"
                },
                "password": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/api/users/forgotPsw": {
      "post": {
        "tags": [
          "User-SignUp & Login/out"
        ],
        "description": "validate",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "email": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/users/member/myProfile": {
      "get": {
        "tags": [
          "User-Member"
        ],
        "description": "validate",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/users/member/myProfile/updatePsw": {
      "patch": {
        "tags": [
          "User-Member"
        ],
        "description": "validate",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "newPassword": {
                  "example": "any"
                },
                "confirmNewPassword": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/users/member/cart": {
      "get": {
        "tags": [
          "User-Member:cart"
        ],
        "description": "validate",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "post": {
        "tags": [
          "User-Member:cart"
        ],
        "description": "post an cart",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "productId": {
                  "example": "any"
                },
                "quantity": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/users/member/cart/{productId}": {
      "patch": {
        "tags": [
          "User-Member:cart"
        ],
        "description": "validate",
        "parameters": [
          {
            "name": "productId",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "quantity": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/users/member/cart/{id}": {
      "delete": {
        "tags": [
          "User-Member:cart"
        ],
        "description": "validate",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        }
      }
    },
    "/api/users/member/orders": {
      "get": {
        "tags": [
          "User-Member:orders"
        ],
        "description": "validate",
        "responses": {
          "default": {
            "description": ""
          }
        }
      }
    },
    "/api/users/member/logout": {
      "post": {
        "tags": [
          "User-Member"
        ],
        "description": "validate",
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/admin/products/": {
      "get": {
        "tags": [
          "Admin-Products"
        ],
        "description": "Get all Products",
        "responses": {
          "200": {
            "description": "got all products successfully.",
            "schema": {
              "type": "array",
              "items": {
                "type": "object",
                "properties": {
                  "title": {
                    "type": "string",
                    "example": "test"
                  },
                  "complete": {
                    "type": "boolean",
                    "example": true
                  },
                  "id": {
                    "type": "string",
                    "example": "202405251645"
                  }
                }
              },
              "xml": {
                "name": "main"
              }
            }
          }
        }
      }
    },
    "/api/admin/products/{productId}": {
      "get": {
        "tags": [
          "Admin-Products"
        ],
        "description": "Get an Products by id",
        "parameters": [
          {
            "name": "productId",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/admin/products/uploadProduct": {
      "post": {
        "tags": [
          "Admin-Products"
        ],
        "description": "upload new product",
        "parameters": [
          {
            "name": "obj",
            "in": "body",
            "description": "add new product.",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "productName": {
                  "type": "string",
                  "example": "Fish"
                },
                "type": {
                  "type": "string",
                  "example": "Fish"
                },
                "order": {
                  "type": "string",
                  "example": "pre-order"
                },
                "price": {
                  "type": "string",
                  "example": "25"
                }
              }
            }
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        }
      }
    },
    "/api/admin/products/updateProduct/{id}": {
      "patch": {
        "tags": [
          "Admin-Products"
        ],
        "description": "update Products by id",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "productName": {
                  "example": "any"
                },
                "type": {
                  "example": "any"
                },
                "order": {
                  "example": "any"
                },
                "price": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/admin/products/deleteProduct/{id}": {
      "delete": {
        "tags": [
          "Admin-Products"
        ],
        "description": "delete Products by id",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/admin/users/": {
      "get": {
        "tags": [
          "Admin-Users"
        ],
        "description": "get all users",
        "responses": {
          "200": {
            "description": "User registered successfully.",
            "schema": {
              "type": "object",
              "properties": {},
              "xml": {
                "name": "main"
              }
            }
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/admin/users/{id}": {
      "get": {
        "tags": [
          "Admin-Users"
        ],
        "description": "get a users bu id",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/admin/users/orders": {
      "post": {
        "tags": [
          "Admin-Orders"
        ],
        "description": "get all orders",
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    }
  }
}